* default makefile target now also rebuilds lhs2TeX.{sty,fmt} * merged stuff from other projects into lhs2TeX.{sty,fmt}:   --  invisible/visible comments   --  save/restore   --  redefine alignment of columns
The basic design principle that I tried (and will try) to follow is
to keep the behaviour of old math style as much as possible unchanged.
Therefore I refrained from the urge to merge the changes into math
style as well. But I don't have the time to test math style, so math
style will not get any new features. Everything that's new will thus
*only* be available in poly style.


git-svn-id: https://subversion.cs.uu.nl/repos/staff.andres.lhs2tex/trunk@20 d54a9f6e-e2dc-0310-8252-bfaaff2a01be