* license changes
Explanation:

This is a (possibly preliminary) reaction to a mail from
Nils Anders Danielsson (NAD), dating back to May 10, 2004. 
In this mail, NAD points out that lhs2TeX prelude files
lhs2TeX.sty and lhs2TeX.fmt are literally included into
files processed by lhs2TeX, causing the resulting .tex
source file to be forced under the GPL if the prelude files
are.

Forcing .tex sources to be GPL when lhs2TeX is used is
not only undesirable, but can even be in conflict with 
license conditions that publishing companies impose on
submitted documents.

I have therefore performed the following changes:

(1) the GPL-2 is now under COPYING, no longer LICENSE
(2) the file LICENSE contains a short explanation stating
    that the program is under GPL, but stating also that
    there is an exception for specific files, and these
    specific files may be literally included into other
    documents, no matter what their license is
(3) the files lhs2TeX.sty, lhs2TeX.fmt, as well as the
    .fmt files in the Library directory now contain a 
    notice in their header stating this exception



git-svn-id: https://subversion.cs.uu.nl/repos/staff.andres.lhs2tex/lhs2TeX/trunk@106 d54a9f6e-e2dc-0310-8252-bfaaff2a01be